[FEATURE] add endpoint to retrieve metainf of jobConfirmationPdf record
This adds a minimal REST endpoint to get meta information of stored
jobconfirmation pdf. In JobConfirmationPdf record the file field was
annotated with @JsonIgnore, so the data is not transfered on every
get request. Added get method to JobConfirmationStorageService and
changed the return value of store to UUID, which saves traffic.

API Contract:

    Request
    URI: /api/jobconfirmations/{uuid}
    HTTP Verb: GET
    Body: None

    Response:
    HTTP Status:
    200 OK if job confirmation was successfully retrieved
    404 NOT FOUND if the job confirmation cannot be found
    Response Body Type: JSON
    Example Response Body:
    {
    "id": "b894e613-618d-42f0-89f1-067efd29dd46",
    "filename": "dummy.pdf"
    }

Closes #3